+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.8.1
+
+- Support `servant-0.14`
+- Add `Lucid.Servant` with `safeHref_` function and its variants.
+
 # 0.8
 
 - Only single `MimeRender HTML a` instance. `lucid-2.9.8` has `ToHtml (HtmlT m a)` instance.